Add -force flag, and decompile NameObjArchiveListCollector

+++ include/MR/StringUtil.h
@@ -6,11 +6,11 @@
 namespace MR
 {
     void addNumberFontTag(wchar_t *, s32);
-    void removeExtensionString(char *, u64, const char *);
+    void removeExtensionString(char *, u32, const char *);
     void makeRaceBestTimeString(wchar_t *, s32);
     void makeRaceCurrentTimeString(wchar_t *);
-    void copyString(char *, const char *, u64);
-    void copyString(wchar_t *, const wchar_t *, u64);
+    void copyString(char *, const char *, u32);
+    void copyString(wchar_t *, const wchar_t *, u32);
 };
 
 #endif // STRINGUTIL_H
